One new example component using readmultiport, simle implementation\n New exampleLib dir layout\n Fixed some memory leaks in core\n Som more doxy comments and some cleanup

+++ HOPSAN++/HopsanCLI/main.cpp
@@ -58,7 +58,7 @@ void printSystemParams(ComponentSystem* pSystem)
     vector<string> names, values, units, descriptions, types;
     pSystem->getParameters(names,values,descriptions,units,types);
 
-    for (int i=0; i<names.size(); ++i)
+    for (size_t i=0; i<names.size(); ++i)
     {
         cout << " SysParam: " << names[i] << "=" << values[i];
     }
